# Lab Work №4/docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "OpenSurvey - Система для проведения онлайн-опросов",
      "description": "ASP.NET Core Web API приложение для проведения онлайн-опросов. В документации описана часть API, связанная с опросами и ответами на опросы",
      "version": "v1"
    },
    "paths": {
      "/surveys": {
        "get": {
          "tags": [
            "Survey"
          ],
          "summary": "Получить список опросов",
          "description": "Возвращает список опросов в соответствии с указанным размером и отступом. В будущем в теле запроса также будут передаваться опции сортировки и фильтрации",
          "parameters": [
            {
              "name": "count",
              "in": "query",
              "description": "Количество опросов в возвращаемом списке",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 20
              }
            },
            {
              "name": "offset",
              "in": "query",
              "description": "Отступ возвращаемого списка опросов",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Запрос успешно выполнен",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Survey"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Survey"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Survey"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Неверно указаны параметры запроса"
            }
          }
        },
        "post": {
          "tags": [
            "Survey"
          ],
          "summary": "Создать и опубликовать опрос",
          "description": "Размещает новый опрос в системе. Для опроса генерируется GUID",
          "requestBody": {
            "description": "Опрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Запрос успешно выполнен",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Survey"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Survey"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Survey"
                  }
                }
              }
            },
            "400": {
              "description": "Ошибка в теле запроса"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Survey": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "nullable": true,
              "readOnly": true
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "questions": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SurveyQuestion"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "SurveyQuestion": {
          "type": "object",
          "properties": {
            "publicationId": {
              "type": "string",
              "nullable": true
            },
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "text": {
              "type": "string",
              "nullable": true
            },
            "questionType": {
              "$ref": "#/components/schemas/SurveyQuestionType"
            },
            "options": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "SurveyQuestionType": {
          "enum": [
            0,
            1,
            2,
            3
          ],
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }